requirements.txt
requests==2.31.0
python-dotenv==1.0.0
cryptography==41.0.7
websockets>=9.0,<12.0
schedule==1.2.0
colorama==0.4.6
tabulate==0.9.0
solana==0.30.2
solders==0.18.1
anchorpy==0.18.0
base58==2.1.1
ledgerblue==0.1.48
hidapi==0.14.0 